+++ .dockerignore
@@ -19,8 +19,8 @@ coverage.xml
 .pytest_cache
 .hypothesis
 
-module/tests
-module/conf/const_dev.py
+src/module/tests
+src/module/conf/const_dev.py
 config/bangumi.json/config/bangumi.json
 /docs
 /.github
