+++ .github/workflows/docker.yml
@@ -14,7 +14,7 @@ jobs:
         uses: actions/checkout@v3
       - name: Create Version info
         run: |
-          echo "version='$GITHUB_REF_NAME'" > module/conf/version.py
+          echo "version='$GITHUB_REF_NAME'" > src/module/__version__.py
       - name: Docker meta
         id: meta
         uses: docker/metadata-action@v4
