+++ .github/workflows/unittest.yml
@@ -17,17 +17,18 @@ jobs:
 
     steps:
     - uses: actions/checkout@v3
-    - name: Set up Python 3.10
+    - name: Set up Python 3.11
       uses: actions/setup-python@v3
       with:
-        python-version: "3.10"
+        python-version: "3.11"
 
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
         pip install pytest
-        mkdir config
 
     - name: Test
-      run: pytest test/test_raw_parser.py
+      run: |
+        cd src
+        pytest test/test_raw_parser.py
