--- module/core/__init__.py
@@ -1,5 +1,0 @@
-from .download_client import DownloadClient
-from .eps_complete import FullSeasonGet
-from .renamer import Renamer
-from .rss_analyser import RSSAnalyser
-from .api_func import APIProcess
