+++ src/module/conf/__init__.py
@@ -1,5 +1,5 @@
-from .config import settings, VERSION
 from .log import setup_logger, LOG_PATH
+from .config import settings, VERSION
 
 
 TMDB_API = "32b19d6a05b512190a056fa4e747cbbc"
