+++ src/module/manager/__init__.py
@@ -1,0 +1,2 @@
+from .eps_complete import FullSeasonGet
+from .renamer import Renamer
