+++ src/module/manager/renamer.py
@@ -1,14 +1,13 @@
 import logging
 import os.path
 import re
-import os.path
 from pathlib import PurePath, PureWindowsPath
 
-from .download_client import DownloadClient
+from module.core.download_client import DownloadClient
 
 from module.conf import settings
 from module.parser import TitleParser
-from ..network import PostNotification, ServerChanNotification
+from module.network import PostNotification, ServerChanNotification
 
 logger = logging.getLogger(__name__)
 
@@ -54,7 +53,6 @@ class Renamer:
         return path_name, season, folder_name, suffix, download_path
 
     def run(self):
-        notification = ServerChanNotification()
         recent_info, torrent_count = self.get_torrent_info()
         rename_count = 0
         for info in recent_info:
@@ -67,10 +65,9 @@ class Renamer:
                 try:
                     new_name = self._renamer.download_parser(name, folder_name, season, suffix, settings.bangumi_manage.rename_method)
                     if path_name != new_name:
-                        old_name = info.content_path.replace(info.save_path, "")
+                        old_name = os.path.basename(info.content_path)
                         self.client.rename_torrent_file(torrent_hash, new_name, old_name, new_name)
                         rename_count += 1
-                        notification.send_msg(f"《{name[:10]}》缓存成功", f"[Auto Bangumi]《{name}》缓存成功")
                     else:
                         continue
                 except Exception as e:
