+++ src/module/rss/__init__.py
@@ -1,0 +1,1 @@
+from .rss_analyser import RSSAnalyser
